+++ libs/base_core/src/num/geom/metric/orientation.rs
@@ -1,4 +1,4 @@
-// devela::num::geom::metric::direction
+// devela_base_core::num::geom::metric::direction
 //
 //! Defines [`Orientation`].
 //