+++ libs/base_core/src/num/geom/metric/stride.rs
@@ -1,4 +1,4 @@
-// devela::num::geom::metric::stride
+// devela_base_core::num::geom::metric::stride
 //
 //! Defines [`Stride`].
 //