+++ libs/base_core/src/num/mod.rs
@@ -12,21 +12,33 @@ mod logic; // ConstBool, False, True, const_bool!
 mod ord;
 
 pub mod error; // error types
+pub mod geom;  // (geometry)
 pub mod niche; // NonZero*, NonZero*, NonValue*|NonExtreme*, ne!, nz!
 pub mod quant; // Cycle*, Interval, interval!, Ratio
 
 crate::structural_mods! { //_mods, _pub_mods
     _mods {
         pub use super::{
-            cast::_all::*, float::_all::*, int::_all::*, logic::_all::*, ord::_all::*,
+            cast::_all::*,
+            float::_all::*,
+            int::_all::*,
+            logic::_all::*,
+            ord::_all::*,
         };
     }
     _pub_mods {
         #[doc(inline)]
         pub use super::error::*;
-        pub use super::{niche::_all::*, quant::_all::*};
+        pub use super::{
+            geom::_all::*,
+            niche::_all::*,
+            quant::_all::*,
+        };
     }
     _workspace_internals {
-        pub use super::niche::_workspace_internals::*;
+        pub use super::{
+            geom::_workspace_internals::*,
+            niche::_workspace_internals::*,
+        };
     }
 }