+++ src/num/geom/mod.rs
@@ -1,6 +1,6 @@
 // devela::num::geom
 //
-//! Geometric types and operations, spatial constructs and analysis.
+#![doc = crate::_DOC_NUM_GEOM!()]
 #![doc = crate::_doc!(modules: crate::num; geom: linear, metric, shape)]
 //
 
@@ -13,7 +13,7 @@ pub mod linear; // Vector*, Matrix*
 #[cfg_attr(nightly_doc, doc(cfg(feature = "shape")))]
 pub mod shape; // Point, …
 
-crate::structural_mods! { // _pub_mods, _crate_internals
+crate::structural_mods! { // _pub_mods
     _pub_mods {
         pub use super::metric::_all::*;
 
@@ -21,8 +21,5 @@ crate::structural_mods! { // _pub_mods, _crate_internals
         pub use super::linear::_all::*;
         #[cfg(feature = "shape")]
         pub use super::shape::_all::*;
-    }
-    _crate_internals {
-        pub(crate) use super::metric::_crate_internals::*;
     }
 }