+++ src/num/mod.rs
@@ -54,7 +54,6 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     }
     _crate_internals {
         pub(crate) use super::{
-            geom::_crate_internals::*,
             rand::_crate_internals::*,
         };
     }